fix(metrics_extraction): user_misery with default threshold is not on-demand
A customer is using the big number widget without a query and user_misery as the aggregate. We're expecting it to be on-demand, however, this is already supported by standard metrics.

This changes the behaviour for `user_misery` to say that it is not supported by on-demand when it uses the default threshold.

Notes:

* We preserve the behaviour of apdex(300) to prevent regressing alerts
* We add partial support for `count_miserable(field, threshold)` even though we do not yet support it for on-demand metrics
  * I made the incorrect decision of not supporting it [1]

[1] https://github.com/getsentry/sentry/issues/58233